## refactor(ecs/factory/player): modulariza lógica de configuración, carga de sprites y creación de entidad
### Cambios realizados:
- **config.py**: extraída la lectura de `players.json` y definidas constantes como `ORIGINAL_SPRITE_SIZE`, `PLAYER_STATS`, `DEFAULT_CLASS`, etc.
- **sprite_loader.py**: movidas y exportadas como públicas las funciones `_load_and_scale_sprites`, `_extract_initial_frame` y `_build_animator_map` como `load_and_scale_sprites`, `extract_initial_frame`, `build_animator_map`.
- **player_factory.py**: limpia la lógica inline de JSON y helpers, ahora enfocado solo en la creación de la entidad ECS. Actualizados los imports a los nuevos módulos.

### Beneficios:
- Aislamiento de la configuración.
- Encapsulamiento claro de la lógica de sprites.
- Factory más simple y orientado a su propósito.
